[LIB-62] Move shader to commo folder
