+++ MNE/disp3D/disp3d.qrc
@@ -1,13 +1,13 @@
 <RCC>
     <qresource prefix="/">
-        <file>model/materials/shaders/gl3/light.inc.frag</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha.frag</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha.geom</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha.vert</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha.tcs</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha_pn_traingles.tes</file>
-        <file>model/materials/shaders/gl3/pervertexphongalpha_simple.tes</file>
-        <file>model/materials/shaders/gl3/networkEdge.frag</file>
-        <file>model/materials/shaders/gl3/networkEdge.vert</file>
+        <file>model/common/shaders/gl3/light.inc.frag</file>
+        <file>model/common/shaders/gl3/networkEdge.frag</file>
+        <file>model/common/shaders/gl3/networkEdge.vert</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha.frag</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha.geom</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha.tcs</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha.vert</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha_pn_traingles.tes</file>
+        <file>model/common/shaders/gl3/pervertexphongalpha_simple.tes</file>
     </qresource>
 </RCC>
