+++ MNE/disp3D/model/common/renderable3Dentity.cpp
@@ -39,7 +39,7 @@
 //=============================================================================================================
 
 #include "renderable3Dentity.h"
-#include "../materials/shadermaterial.h"
+#include "shadermaterial.h"
 
 
 //*************************************************************************************************************
