+++ MNE/disp3D/model/common/shaders/gl3/pervertexphongalpha.vert
